requirements.txt
Flask
requests
pillow
fastapi
uvicorn
transformers
torch